Enable source maps in Obsidian 0.14+
In Obsidian 14 and above, sourcemaps are stripped from plugins
at load time unless the "debug plugin startup" setting is on.
Hot Reload now automatically switches that option on when
reloading a plugin, so you get sourcemaps while developing,
without needing to keep the debug setting on all the time
(or sourcemaps in memory all the time).